% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute}
\alias{impute}
\title{\code{impute}}
\usage{
impute(x, s, posit_y, posit_x, w)
}
\arguments{
\item{posit_y}{an integer - a position of dependent variable}

\item{posit_x}{an integer vector - positions of independent variables}

\item{w}{a numeric vector - a weighting variable - only positive values}

\item{dt}{data.table - a data.table with variables}

\item{model}{a character - a posibble options ("lda","lm_pred","lm_bayes","lm_noise")}
}
\value{
load data in a data.table format
}
\description{
\code{impute}
}
\examples{
\dontrun{
   fars_read(filename)
}

}
\seealso{
\code{\link{impute}}
}
