% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_N}
\alias{impute_N}
\title{\code{impute_N}}
\usage{
impute_N(x, s, posit_y, posit_x, times, w = 0L)
}
\arguments{
\item{posit_y}{an integer - a position of dependent variable}

\item{posit_x}{an integer vector - positions of independent variables}

\item{times}{an integer - a number of multiple imputations}

\item{w}{a numeric vector - a weighting variable - only positive values}

\item{dt}{data.table - a data.table with variables}

\item{model}{a character - a posibble options ("lm_bayes","lm_noise")}
}
\value{
load data in a data.table format
}
\description{
\code{impute_N}
}
\examples{
\dontrun{
   fars_read(filename)
}

}
\seealso{
\code{\link{impute}}
}
