\name{miceFast-package}
\alias{miceFast-package}
\docType{package}
\title{
miceFast package for foast multiple imputations.
}
\description{
Fast imputations under the object-oriented programming paradigm.
There was used quantitative models with a closed-form solution. Thus package is based on linear algebra operations.
The biggest improvement in time performance could be achieve for a calculation where a grouping variable have to be used.
}
\details{
read vignette for additional information
}
\author{
Maciej Nasinski
}
\references{
.
}
\keyword{ package }
\seealso{.}
\examples{
  \dontrun{
.
  }
}
