% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_fill_NA.R
\name{naive_fill_NA}
\alias{naive_fill_NA}
\title{\code{naive_fill_NA} function for the imputations purpose.}
\usage{
naive_fill_NA(x)
}
\arguments{
\item{x}{a numeric matrix or data.frame/data.table (factor/character/numeric/logical)  - variables}
}
\value{
load imputations in a similar to the input type
}
\description{
Regular imputations to fill the missing data.
Non missing independent variables are used to approximate a missing observations for a dependent variable.
Quantitative models were built under Rcpp packages and the C++ library Armadillo.
}
\note{
There is assumed that users add the intercept by their own.
The miceFast module provides the most efficient environment, the second recommended option is to use data.table and the numeric matrix data type.
The lda model is assessed only if there are more than 15 complete observations
and for the lms models if number of independent variables is smaller than number of observations.
}
\examples{
\dontrun{
library(miceFast)
data(air_miss)
naive_fill_NA(air_miss)
}

}
\seealso{
\code{\link{fill_NA}} \code{\link{fill_NA_N}}  \code{\link{VIF}}
}
