% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neibo}
\alias{neibo}
\title{Finding in random manner one of the k closest points in a certain vector
for each value in a second vector}
\usage{
neibo(y, miss, k)
}
\arguments{
\item{y}{numeric vector values to be looked up}

\item{miss}{numeric vector values to be looked for}

\item{k}{integer number of nearest neighbours to sample from}
}
\value{
a numeric vector
}
\description{
This function uses pre-sorting of y and binary search to find
  one of the k closest values for each miss.
}
