% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vifs}
\alias{vifs}
\title{\code{vif}}
\usage{
vifs(x, posit_y, posit_x)
}
\arguments{
\item{x}{a numeric matrix - a numeric matrix with variables}

\item{posit_y}{an integer - a position of dependent variable}

\item{posit_x}{an integer vector - positions of independent variables}
}
\value{
load a numeric vector with VIF for all variables matching with a provided posit_x
}
\description{
\code{vif}
}
\examples{
\dontrun{
   impute(filename)
}

}
\seealso{
\code{\link{impute_N}}
}
