\name{miceFast-package}
\alias{miceFast-package}
\docType{package}
\title{
miceFast package for fast multiple imputations.
}
\description{
  Fast imputations under the object-oriented programming paradigm.
	There was used quantitative models with a closed-form solution. Thus package is based on linear algebra operations.
	The biggest improvement in time performance could be achieve for a calculation where a grouping variable have to be used.
	A single evaluation of a quantitative model for the multiple imputations is another major enhancement.
	Moreover there are offered a few functions built to work with popular R packages such as 'data.table'.
}
\details{
Please read the vignette for additional information
}
\author{
Maciej Nasinski
}
\references{
\url{https://github.com/Polkas/miceFast}
}
